+++ CHANGES.rst
@@ -23,6 +23,8 @@ Changelog
 - Fix getting container for some cases.
   [pbauer]
 
+- Fix use in Plone 4.3 without dexterity, zc.relation or plone.app.contenttypes.
+  [pbauer]
 
 1.0 (2021-04-27)
 ----------------